MNT: move table models into tables/
