+++ squirrel/tables/snapshot_comparison_table.py
