+++ squirrel/tables/snapshot_table.py
