+++ squirrel/widgets/squirrel_table_view.py
